migrate DrinkList Buy dialog into separate module
